refactor: keep testutil and reflectutil as internal
