+++ core/scanner_test.go
@@ -11,9 +11,9 @@ import (
 	_ "github.com/go-sql-driver/mysql"
 	_ "github.com/jackc/pgx/v5/stdlib"
 	"github.com/rfberaldo/sqlz/core"
+	"github.com/rfberaldo/sqlz/internal/testutil"
+	"github.com/rfberaldo/sqlz/internal/testutil/mock"
 	"github.com/rfberaldo/sqlz/parser"
-	"github.com/rfberaldo/sqlz/testutil"
-	"github.com/rfberaldo/sqlz/testutil/mock"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
