+++ sqlz_test.go
@@ -8,8 +8,8 @@ import (
 	"time"
 
 	"github.com/rfberaldo/sqlz"
+	"github.com/rfberaldo/sqlz/internal/testutil"
 	"github.com/rfberaldo/sqlz/parser"
-	"github.com/rfberaldo/sqlz/testutil"
 	"github.com/stretchr/testify/assert"
 
 	_ "github.com/go-sql-driver/mysql"
